+++ lizard_progress/models.py
@@ -534,29 +534,9 @@ class Project(ProjectActivityMixin, models.Model):
             return self.created_at
 
     def archive(self):
-        """Archive a project.
-
-        This has a side effect for Projects with measurements with measurement
-        types that have the delete_on_archive field set: those measurements
-        will be deleted, both in db and on disk. The motivation for this is
-        that we want to remove 'attachment' measurement files, e.g., all media
-        files belonging to a ribx, but not the ribx itself.
-        """
-        logger.info("Archiving project %s", self)
-        # This query returns all Measurements that (1) belong to a Project,
-        # (2) that have a parent Measurement, which entails that it is an
-        # attachment (e.g., media files belonging to a Ribx), and (3) have
-        # a measurement type that can be deleted.
-        measurements = Measurement.objects.filter(
-            location__activity__project=self,
-            parent__isnull=False,
-            location__activity__measurement_type__delete_on_archive=True)
-        logger.debug("Deleting measurements: %s", measurements)
-        for m in measurements:
-            m.delete(notify=False, deleted_by_contractor=False,
-                     set_completeness=False)
-        self.is_archived = True
-        self.save()
+        """Archive the project using a Celery task."""
+        from . import tasks
+        tasks.archive_task.delay(self.id)
 
     def activate(self):
         self.is_archived = False
@@ -1132,9 +1112,23 @@ class ExpectedAttachment(models.Model):
                 filename__iexact=filename)
         except cls.DoesNotExist:
             return
-
-        expected_attachment.uploaded = False
-        expected_attachment.save()
+        except cls.MultipleObjectsReturned:
+            # We shouldn't normally get here. This exception is to remedy a
+            # situation where case sensitive  uploads where permitted, i.e.,
+            # same filenames with lower or upper case characters. In July 2015
+            # a commit was introduced that made it not possible anymore (I
+            # think), but the projects that were created before that commit
+            # still have same filenames that can have different cases.
+            expected_attachments = cls.objects.distinct().filter(
+                measurements__location__activity=activity,
+                filename__iexact=filename)
+            for att in expected_attachments:
+                att.uploaded = False
+                att.save()
+            return
+        else:
+            expected_attachment.uploaded = False
+            expected_attachment.save()
 
     class Meta:
         ordering = ('uploaded', 'filename', )